Move JSON-specific assert method for reuse [HZ-1742]
This moves `assertJsonRowsAnyOrder` and related helpers from
`JsonSqlAggregateTest` to `SqlJsonTestSupport` for reuse in other tests.
